feat: add OpenTelemetry trace context to telemetry export
Rollback trigger: telemetry serialization errors
